server: Reworked error handling in mail and rcpt.
The Erro handler used a defered statement to print the correct error
message. This was replaced by explicit calls to the ErrCmd helper.

The Error code was also corrected from 553 to 501 for all cases where
the paramter error was meant and not the mailbox syntax error.

This change should make the code a little more readable.